cci_image/requirements.txt
werkzeug==2.3.*
waitress==2.1.*
esdk-obs-python==3.22.*
python-dotenv==1.0.*
